requirements.txt
python-telegram-bot[job-queue]>=20.0
groq>=1.12.0>=0.4.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
nest-asyncio>=1.6.0
pytz>=2024.1
matplotlib>=3.8.0
